abstracted store and locale setup into Tool component
