+++ src/Container.js
@@ -6,30 +6,22 @@ import HTML5Backend from 'react-dnd-html5-backend';
 import WordBankArea from './components/WordBankArea';
 import DropBoxArea from './components/DropBoxArea';
 import isEqual from 'lodash/isEqual';
-import {configureStore, Provider} from './state/store';
-import {loadLocalization} from './state/actions/locale';
-import {setActiveLanguage} from 'react-localize-redux';
 import path from 'path-extra';
+import Tool, {createConnect} from './Tool';
 
+const STORE_KEY = 'wordAlignment';
 
+/**
+ * The custom connect HOC for this tool
+ */
+exports.connect = createConnect(STORE_KEY);
+
+/**
+ * The base container for this tool
+ */
 class Container extends Component {
 
-  constructor(props) {
-    super(props);
-    this.state = {
-      store: null
-    };
-  }
-
   componentWillMount() {
-    const {appLanguage} = this.props;
-    const store = configureStore();
-    this.setState({
-      store
-    });
-    const localeDir = path.join(__dirname, '../locale');
-    store.dispatch(loadLocalization(localeDir, appLanguage));
-
     // current panes persisted in the scripture pane settings.
     const {ScripturePane} = this.props.settingsReducer.toolsSettings;
     let panes = [];
@@ -53,12 +45,6 @@ class Container extends Component {
       let page = document.getElementById("DropBoxArea");
       if (page) page.scrollTop = 0;
     }
-
-    // stay in sync with the application language
-    if(nextProps.appLanguage !== this.props.appLanguage) {
-      const {store} = this.state;
-      store.dispatch(setActiveLanguage(nextProps.appLanguage));
-    }
   }
 
   render() {
@@ -70,10 +56,13 @@ class Container extends Component {
       scripturePane = <ScripturePane {...this.props} />;
     }
 
-    const {store} = this.state;
+    const {appLanguage} = this.props;
+    const localeDir = path.join(__dirname, '../locale');
 
     return (
-      <Provider store={store}>
+      <Tool appLanguage={appLanguage}
+            storeKey={STORE_KEY}
+            localeDir={localeDir}>
         <div style={{ display: 'flex', width: '100%', height: '100%' }}>
           <WordBankArea {...this.props} />
           <div style={{ flex: 0.8, width: '100%', height: '100%', paddingBottom: '150px' }}>
@@ -81,7 +70,7 @@ class Container extends Component {
             <DropBoxArea {...this.props} />
           </div>
         </div>
-      </Provider>
+      </Tool>
     );
   }
 }
